// src/smart-tags/domain/enums/smart-tag-type.enum.ts
export enum SmartTagType {
	INTEREST = "INTEREST", // Interés del cliente
	CUSTOMER_TYPE = "CUSTOMER_TYPE", // Tipo de servicio
	BUYER_TYPE = "BUYER_TYPE", // Tipo de comprador
	PROBLEM = "PROBLEM", // Situación actual del cliente
	FEELING = "FEELING", // Vínculo o relación personal
	NEED = "NEED", // Motivación o necesidad
	OBJECTIVE = "OBJECTIVE", // Valor para el negocio
	GOAL = "GOAL", // Meta
	ACTION = "ACTION", // Hábitos o comportamiento
	SOLUTION = "SOLUTION", // Solución
	RESULT = "RESULT", // Resultado
}